# api/openapi.yaml
openapi: 3.0.0
info:
  title: API de Categorías
  version: 1.0.0
  description: Una API para gestionar categorías.
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo local
paths:
...
  /categories:
    get:
      tags:
        - categories
      summary: Lista todas las categorías
      responses:
        '200':
          description: Una lista de categorías
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'

    post:
      tags:
        - categories
      security:
        - bearerAuth: []
      summary: Crea una nueva categoría
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: El nombre de la categoría
                images:
                  type: string
                  format: binary
                  description: La imagen de la categoría
              required:
                - name
                - images
      responses:
        '201':
          description: Categoría creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Solicitud inválida
  /categories/{categoryId}:
    patch:
      tags:
        - categories
      security:
        - bearerAuth: []
      summary: Actualiza parcialmente una categoría específica por su ID
      parameters:
        - in: path
          name: categoryId
          required: true
          schema:
            type: number
            description: El ID de la categoría a actualizar
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: El nombre de la categoría
                images:
                  type: string
                  format: binary
                  description: La imagen de la categoría
              description: Propiedades de la categoría para actualizar
      responses:
        '200':
          description: Categoría actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Datos de entrada inválidos
        '404':
          description: Categoría no encontrada

    delete:
      tags:
        - categories
      security:
        - bearerAuth: []
      summary: Elimina una categoría específica por su ID
      parameters:
        - in: path
          name: categoryId
          required: true
          schema:
            type: number
            description: El ID de la categoría a eliminar
      responses:
        '200':
          description: Categoría eliminada exitosamente
        '404':
          description: Categoría no encontrada
  /client-assistance/text:
    post:
      tags:
        - client-assistance
      summary: Crea una nueva asistencia de cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sessionId:
                  type: string
                  description: El ID de la sesión del cliente
                content:
                  type: string
                  description: El contenido de la asistencia
              required:
                - sessionId
                - content
      responses:
        '201':
          description: Respondido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssistanceAction'
        '400':
          description: Solicitud inválida
...
  /ingredients:
    get:
      tags:
        - ingredients
      summary: Lista todos los ingredientes
      responses:
        '200':
          description: Una lista de ingredientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ingredient'

    post:
      tags:
        - ingredients
      security:
        - bearerAuth: []
      summary: Crea un nuevo ingrediente
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: El nombre del ingrediente
                images:
                  type: string
                  format: binary
                  description: La imagen del ingrediente
              required:
                - name
                - images
      responses:
        '201':
          description: Ingrediente creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
        '400':
          description: Solicitud inválida
  /ingredients/{ingredientId}:
    patch:
      tags:
        - ingredients
      security:
        - bearerAuth: []
      summary: Actualiza parcialmente un ingrediente específico por su ID
      parameters:
        - in: path
          name: ingredientId
          required: true
          schema:
            type: number
            description: El ID del ingrediente a actualizar
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: El nombre del ingrediente
                images:
                  type: string
                  format: binary
                  description: La imagen del ingrediente
              description: Propiedades del ingrediente para actualizar
      responses:
        '200':
          description: Ingrediente actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
        '400':
          description: Datos de entrada inválidos
        '404':
          description: Ingrediente no encontrado

    delete:
      tags:
        - ingredients
      security:
        - bearerAuth: []
      summary: Elimina una categoría específica por su ID
      parameters:
        - in: path
          name: ingredientId
          required: true
          schema:
            type: number
            description: El ID del ingrediente a eliminar
      responses:
        '200':
          description: Ingrediente eliminado exitosamente
        '404':
          description: Ingrediente no encontrado
  /mercadopago/devices:
    get:
      tags:
        - mercadopago
      security:
        - bearerAuth: []
      summary: Lista todos los dispositivos de MercadoPago
      parameters:
        - in: query
          name: page
          required: true
          schema:
            type: integer
            default: 0
          description: Número de página de la lista de dispositivos de MercadoPago
        - in: query
          name: pageSize
          required: true
          schema:
            type: integer
            default: 10
          description: Cantidad de dispositivos de MercadoPago por página
      responses:
        '200':
          description: Una lista de dispositivos de MercadoPago
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/MercadoPagoDevice'
                  pagination:
                    $ref: '#/components/schemas/PaginationResult'
  /mercadopago/devices/{mpDeviceId}/operating-mode:
    patch:
      tags:
        - mercadopago
      security:
        - bearerAuth: []
      summary: Actualiza el modo de operación de un dispositivo de MercadoPago
      parameters:
        - in: path
          name: mpDeviceId
          required: true
          schema:
            type: string
            description: El ID del dispositivo de MercadoPago
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                operatingMode:
                  type: string
                  enum:
                    - PDV
                    - STANDALONE
                  description: El modo de operación del dispositivo
              required:
                - operatingMode
      responses:
        '200':
          description: Modo de operación actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MercadoPagoDevice'
        '400':
          description: Datos de entrada inválidos
        '404':
          description: Dispositivo de MercadoPago no encontrado
  /mercadopago/devices/{mpDeviceId}/payment-intents:
    post:
      tags:
        - mercadopago
      security:
        - bearerAuth: []
      summary: Crea un nuevo intento de pago en MercadoPago
      parameters:
        - in: path
          name: mpDeviceId
          required: true
          schema:
            type: string
            description: El ID del dispositivo de MercadoPago
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  description: El monto del intento de pago
                ticketNumber:
                  type: string
                  description: El número de ticket del intento de pago
                externalReference:
                  type: string
                  description: El ID externo del intento de pago
              required:
                - amount
                - ticketNumber
                - externalReference
      responses:
        '201':
          description: Intento de pago creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: El ID del intento de pago
        '400':
          description: Datos de entrada inválidos
        '404':
          description: Dispositivo de MercadoPago no encontrado
  /mercadopago/devices/{mpDeviceId}/payment-intents/{mpPaymentIntentId}:
    delete:
      tags:
        - mercadopago
      security:
        - bearerAuth: []
      summary: Elimina un intento de pago en MercadoPago
      parameters:
        - in: path
          name: mpDeviceId
          required: true
          schema:
            type: string
            description: El ID del dispositivo de MercadoPago
        - in: path
          name: mpPaymentIntentId
          required: true
          schema:
            type: string
            description: El ID del intento de pago en MercadoPago
      responses:
        '200':
          description: Intento de pago eliminado exitosamente
        '404':
          description: Intento de pago no encontrado
  /mercadopago/stores:
    get:
      summary: Obtiene todas las tiendas
      operationId: getAllStores
      responses:
        '200':
          description: Lista de tiendas del usuario
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Store'
        '400':
          description: Solicitud inválida (p.ej., mpUserId faltante o inválido)
        '404':
          description: Usuario o tiendas no encontradas
        '500':
          description: Error interno del servidor
    post:
      summary: Crea una nueva tienda en MercadoPago
      operationId: createStore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mpUserId:
                  type: number
                  description: ID del usuario de MercadoPago
                storeId:
                  type: string
                  description: ID de la tienda en MercadoPago
                externalId:
                  type: string
                  description: ID externo para la tienda
              required:
                - mpUserId
                - storeId
                - externalId
      responses:
        '201':
          description: Tienda creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Store'
        '400':
          description: Solicitud inválida (p.ej., parámetros faltantes o inválidos)
        '500':
          description: Error interno del servidor
  /mercadopago/stores/{mpStoreId}:
    delete:
      summary: Elimina una tienda específica por su ID
      operationId: deleteStore
      parameters:
        - name: mpStoreId
          in: path
          required: true
          description: El ID de la tienda de MercadoPago
          schema:
            type: string
      responses:
        '200':
          description: Tienda eliminada exitosamente
        '400':
          description: Solicitud inválida (p.ej., mpStoreId faltante o inválido)
        '404':
          description: Tienda no encontrada
        '500':
          description: Error interno del servidor
  /mercadopago-services/{mpUserId}/stores:
    get:
      summary: Obtiene todas las tiendas de un usuario de MercadoPago
      operationId: getAllMercadoPagoStores
      parameters:
        - name: mpUserId
          in: path
          required: true
          description: El ID del usuario de MercadoPago
          schema:
            type: string
      responses:
        '200':
          description: Lista de tiendas del usuario
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Store'
        '400':
          description: Solicitud inválida (p.ej., mpUserId faltante o inválido)
        '404':
          description: Usuario o tiendas no encontradas
        '500':
          description: Error interno del servidor
  /mercadopago/cashiers:
    get:
      summary: Obtiene todas las cajas
      operationId: getAllCashiers
      responses:
        '200':
          description: Lista de cajas (cashiers)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cashier'
        '400':
          description: Solicitud inválida (p.ej., mpStoreId faltante o inválido)
        '404':
          description: Tienda o cajas no encontradas
        '500':
          description: Error interno del servidor
    post:
      summary: Crea una nueva caja (cashier) en MercadoPago
      operationId: createCashier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                storeId:
                  type: string
                  description: ID de la tienda en MercadoPago a la que pertenece la caja
                cashierId:
                  type: number
                  description: ID de la caja en MercadoPago
                externalId:
                  type: string
                  description: ID externo para la caja
              required:
                - storeId
                - cashierId
                - externalId
      responses:
        '201':
          description: Caja creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cashier'
        '400':
          description: Solicitud inválida (p.ej., parámetros faltantes o inválidos)
        '500':
          description: Error interno del servidor
  /mercadopago/cashiers/{mpCashierId}:
    delete:
      summary: Elimina una caja específica por su ID
      operationId: deleteCashier
      parameters:
        - name: mpCashierId
          in: path
          required: true
          description: El ID de la caja de MercadoPago
          schema:
            type: string
      responses:
        '200':
          description: Caja eliminada exitosamente
        '400':
          description: Solicitud inválida (p.ej., mpCashierId faltante o inválido)
        '404':
          description: Caja no encontrada
        '500':
          description: Error interno del servidor
  /mercadopago-services/stores/{mpStoreId}/cashiers:
    get:
      summary: Obtiene todas las cajas de una tienda específica
      operationId: getAllMercadoPagoCashiers
      parameters:
        - name: mpStoreId
          in: path
          required: true
          description: El ID de la tienda de MercadoPago
          schema:
            type: string
      responses:
        '200':
          description: Lista de cajas (cashiers) de la tienda
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cashier'
        '400':
          description: Solicitud inválida (p.ej., mpStoreId faltante o inválido)
        '404':
          description: Tienda o cajas no encontradas
        '500':
          description: Error interno del servidor
  /mercado-pago-webhook:
    post:
      tags:
        - mercado-pago-webhook
      summary: Recibe notificaciones de Mercado Pago
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Notificación recibida
...
  /orders:
    get:
      tags:
        - orders
      summary: Lista todas las órdenes
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum:
              - PAYMENT_PENDING
              - PREPARING
              - READY
              - COMPLETED
              - CANCELED
          description: El estado de las órdenes a filtrar
          allowEmptyValue: true
        - in: query
          name: sortOrder
          schema:
            type: string
            enum:
              - ASC
              - DESC
          description: El orden de las órdenes
          allowEmptyValue: true
        - in: query
          name: page
          schema:
            type: integer
            default: 0
          description: Número de página de la lista de ingresos de mercadería
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 10
          description: Cantidad de ingresos de mercadería por página

      responses:
        '200':
          description: Una lista de órdenes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    post:
      tags:
        - orders
      summary: Crea una nueva orden
      requestBody:
        description: Datos necesarios para crear una nueva orden
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Orden creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Solicitud inválida
        '500':
          description: Error interno del servidor
  /orders/{orderId}:
    get:
      tags:
        - orders
      security:
        - bearerAuth: []
      summary: Obtiene una orden específica por su ID
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: number
          description: El ID de la orden a obtener
      responses:
        '200':
          description: Orden encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Orden no encontrada
    patch:
      tags:
        - orders
      security:
        - bearerAuth: []
      summary: Actualiza parcialmente una orden específica por su ID
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: number
            description: El ID de la orden a actualizar
      requestBody:
        description: Datos necesarios para actualizar una orden
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum:
                    - PAYMENT_PENDING
                    - PREPARING
                    - READY
                    - COMPLETED
                    - CANCELED
                  description: El estado de la orden
              required:
                - status
      responses:
        '200':
          description: Orden actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Datos de entrada inválidos
        '404':
          description: Orden no encontrada
  /orders/{orderId}/pay:
    post:
      tags:
        - orders
      security:
        - bearerAuth: []
      summary: Paga una orden
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: number
          description: El ID de la orden a pagar
      requestBody:
        description: Datos necesarios para pagar una orden
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                method:
                  type: string
                  enum:
                    - cash
                    - card
                  description: El método de pago de la orden
                amount:
                  type: number
                  description: El monto a pagar
              required:
                - method
                - amount
      responses:
        '200':
          description: Orden pagada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Solicitud inválida
        '404':
          description: Orden no encontrada
...
  /orders/{orderId}/pay-mercadopago-dynamic-qr:
    post:
      tags:
        - orders
      security:
        - bearerAuth: []
      summary: Paga una orden con Mercado Pago QR
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: number
          description: El ID de la orden a pagar
      responses:
        '200':
          description: Orden pagada exitosamente
        '400':
          description: Solicitud inválida
        '404':
          description: Orden no encontrada
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT 
  schemas:
...
    AssistanceAction:
      type: object
...
    Category:
      type: object
      properties:
        id:
          type: number
          description: El ID único de la categoría
        name:
          type: string
          description: El nombre de la categoría
        image:
          type: string
          description: La imagen de la categoría
      required:
        - id
        - name
        - image
...
    Ingredient:
      type: object
      properties:
        id:
          type: number
          description: El ID único del ingrediente
        name:
          type: string
          description: El nombre del ingrediente
        image:
          type: string
          description: La imagen del ingrediente
      required:
        - id
        - name
        - image
    MercadoPagoDevice:
      type: object
      properties:
        id:
          type: number
          description: El ID único del dispositivo
        posId:
          type: number
          description: El ID del punto de venta
        storeId:
          type: number
          description: El ID de la tienda
        externalPosId:
          type: string
          description: El ID externo del punto de venta
        operatingMode:
          type: string
          enum:
            - TEST
            - PRODUCTION
          description: El modo de operación del dispositivo
...